# Introduce growInRange to reduce array overallocation