In cases where we know there is an upper limit to the potential size of an array, we can use `growInRange` to avoid allocating beyond that limit.

We address such cases in `DirectoryTaxonomyReader` and `NeighborArray`.

Closes #12839