@benwtrent 2023-11-29
I ran knnPerfTest in Lucene util against this PR. 100k cohere vectors. Flushing was depending on memory usage (Do we check the ram usage of onHeapgraph during indexing to determine when to flush?)

main branch:
```
recall	latency	nDoc	fanout	maxConn	beamWidth index
0.912	 0.77	100000	0	16	100	  48602
```

This branch:
```
recall	latency	nDoc	fanout	maxConn	beamWidth index ms
0.912	 0.75	100000	0	16	100	  165929
```

Indexing is about 3.5x slower with this change. I don't know if its due to the OnHeapGraph memory estimation being slow or the node resizing. I am gonna run a profiler to see whats up.

Good news is that search latency and recall are unchanged. Forced merging time seems about the same as well :).