@benwtrent 2023-11-29
Yeah, `ramBytesUsed` changes here are adding significant overhead to indexing. Here is a cpu profile.

[pr12844-768-100000-wall.jfr.zip](https://github.com/apache/lucene/files/13500964/pr12844-768-100000-wall.jfr.zip)