@stefanvodita 2023-11-29
Thank you for running the benchmarks @benwtrent, you were quicker than I was. I was worried the modified `ramBytesUsed` would be slow. Maybe we can come up with a smarter way to do the estimation.