@stefanvodita 2023-11-30
I thought some more about option 2. It does seem quite tricky. `OnHeapHnswGraph` only knows about `NeighborArrays` being created, but it doesn't know about nodes being added to the arrays - `HnswGraphBuilder` handles that. I don't know this code well, so I could be missing something.

I'm not sure if a more precise estimate is worth pursuing or not. I've pushed the version where we keep memory estimation like it was and I've added @dungba88's suggestions. I'll try to do a few more benchmark runs to see if there is a measurable slow-down and if we can reduce it by increasing `INITIAL_CAPACITY`.