@zhaih 2023-11-30
Yeah I don't think we need to spend too much effort on option2, especially
in this PR, because providing an upper bound on memory usage is good enough
to me. Plus we'll never be able to precisely account how much the memory is
gonna be used.
Maybe leaving a TODO or just mentioning in the comment that this estimation
is not accurate seems ok to me.

On Thu, Nov 30, 2023 at 1:20 PM Stefan Vodita ***@***.***>
wrote:

> I thought some more about option 2. It does seem quite tricky.
> OnHeapHnswGraph only knows about NeighborArrays being created, but it
> doesn't know about nodes being added to the arrays - HnswGraphBuilder
> handles that. I don't know this code well, so I could be missing something.
>
> I'm not sure if a more precise estimate is worth pursuing or not. I've
> pushed the version where we keep memory estimation like it was and I've
> added @dungba88 <https://github.com/dungba88>'s suggestions. I'll try to
> do a few more benchmark runs to see if there is a measurable slow-down and
> if we can reduce it by increasing INITIAL_CAPACITY.
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/apache/lucene/pull/12844#issuecomment-1834578138>, or
> unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AEFSB7ALXIILG7WBJTDKO5DYHD2C7AVCNFSM6AAAAAA7ZUIX5WVHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMYTQMZUGU3TQMJTHA>
> .
> You are receiving this because you were mentioned.Message ID:
> ***@***.***>
>