@stefanvodita 2023-12-01
I did some memory profiling and it doesn't look promising. Let's take initial capacity 100 as an example. Total allocation is 3.41GB compared to 1.29GB on the baseline, and peak heap usage is 946MiB compared to 547MiB on baseline. It's worse for smaller initial capacities and larger capacities devolve into the baseline behavior.

I might try a few runs on a different data-set, but maybe these neighbor arrays tend to be mostly used in the general case.


Baseline: [baseline.jfr.zip](https://github.com/apache/lucene/files/13530935/baseline.jfr.zip)
<img width="1127" alt="image" src="https://github.com/apache/lucene/assets/41467371/a2a09c9c-ef91-4227-a4f8-655404b2cb21">


Candidate: [candidate-100.jfr.zip](https://github.com/apache/lucene/files/13530936/candidate-100.jfr.zip)
<img width="1127" alt="image" src="https://github.com/apache/lucene/assets/41467371/80c2141a-fc6f-4624-9f7a-fb264ad36ef5">