@stefanvodita 2023-12-04
Unless anyone has other ideas, I’ll revert the changes to `NeighborArray` and only keep `growInRange` for `DirectoryTaxonomyReader`. Separately, I will go through other uses of the array growth API and see if `growInRange` is more appropriate.