@stefanvodita 2023-12-09
I reverted the changes to `NeighborArray` and `OnHeapHnswGrap`, since it doesn't look like we're making an improvement there.