@stefanvodita 2023-12-14
Done, thank you @zhaih! I've opened #12941 to replace other uses of the unbounded growth API.